Use new GPT-J chatbot model.

+++ geniusbot/geniusbot_chat.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from transformers import AutoTokenizer, GPTJForCausalLM, pipeline, AutoModelForCausalLM
-from torch.quantization.qconfig import float_qparams_weight_only_qconfig
+from transformers import AutoTokenizer, GPTJForCausalLM, pipeline
 import torch
 import psutil
+import os
+
 
 class ChatBot():
     def __init__(self):
@@ -12,57 +13,66 @@ class ChatBot():
         self.used_memory = psutil.virtual_memory()[3]
         self.free_memory = psutil.virtual_memory()[1]
         self.used_percent = psutil.virtual_memory()[2]
+        self.model = None
+        self.tokenizer = None
+        self.generator = None
+        self.device = 'cuda' if torch.cuda.is_available() else 'cpu'
+        self.memory_required = 50 #GB
+        self.loaded = False
 
     def check_hardware(self):
         self.total_memory = psutil.virtual_memory()[0]
         self.used_memory = psutil.virtual_memory()[3]
         self.free_memory = psutil.virtual_memory()[1]
         self.used_percent = psutil.virtual_memory()[2]
-        print(f"Used RAM: {round(float(self.used_memory/1073741824), 2)} GB\n"
+        print(f"Required RAM: {self.memory_required} GB\n"
+              f"Used RAM: {round(float(self.used_memory/1073741824), 2)} GB\n"
               f"Free RAM: {round(float((self.free_memory)/1073741824), 2)} GB\n"
               f"Total RAM: {round(float(self.total_memory/1073741824), 2)} GB\n"
               f"RAM Utilization Percentage: {round(self.used_percent, 2)}%")
 
     def save_model(self):
-        print("Downloading model")
-        # load fp 16 model
-        model = GPTJForCausalLM.from_pretrained("EleutherAI/gpt-j-6B", revision="float16", torch_dtype=torch.float16, low_cpu_mem_usage=True)
-        # save model with torch.save
-        print("Saving model")
-        torch.save(model, "geniusbot_gptj.pt")
-        print("Model Saved")
+        if os.path.isfile("geniusbot_gptj.pt"):
+            print("Model already downloaded")
+        else:
+            print("Downloading model")
+            # load fp 16 model
+            model = GPTJForCausalLM.from_pretrained("EleutherAI/gpt-j-6B", revision="float16", torch_dtype=torch.float16, low_cpu_mem_usage=True)
+            # save model with torch.save
+            print("Saving model")
+            torch.save(model, "geniusbot_gptj.pt")
+            print("Model Saved")
 
     def load_model(self):
-        if self.free_memory >= 15032385536:
+        if self.free_memory >= (self.memory_required*1073741824):
             print("Loading Model")
-            device = 'cuda' if torch.cuda.is_available() else 'cpu'
-            #torch.device('cpu')
             # load model
-            #model = torch.load("geniusbot_gptj.pt")#.to("cuda",  torch_dtype=torch.float16)
-            #model = torch.load("geniusbot_gptj.pt", map_location=torch.device('cpu'))
-            model = GPTJForCausalLM.from_pretrained("EleutherAI/gpt-j-6B", revision="float16", torch_dtype=torch.float16, low_cpu_mem_usage=True)
-            model.load_state_dict(torch.load("geniusbot_gptj.pt"))
-            model.device(device)
-            # model = Model()
-            # model.load_state_dict(torch.load("geniusbot_gptj.pt", map_location=torch.device('cpu')))
-            # torch.load('tensors.pt', map_location=lambda storage, loc: storage.cuda(1))
+            self.model = GPTJForCausalLM.from_pretrained("EleutherAI/gpt-j-6B", revision="float16", torch_dtype=torch.float16, low_cpu_mem_usage=True)
+            self.model.load_state_dict(torch.load("geniusbot_gptj.pt"))
+            self.model.device(self.device)
             # load tokenizer
-            tokenizer = AutoTokenizer.from_pretrained("EleutherAI/gpt-j-6B")
-            #input_ids = tokenizer(context, return_tensors="pt").input_ids.to("cuda")
+            self.tokenizer = AutoTokenizer.from_pretrained("EleutherAI/gpt-j-6B")
             # create pipeline
-            gen = pipeline("text-generation", model=model, tokenizer=tokenizer)
-
-            # run prediction
-            gen("My Name is GeniusBot")
-            #[{'generated_text': 'My Name is philipp k. and I live just outside of Detroit....
+            self.generator = pipeline("text-generation", model=self.model, tokenizer=self.tokenizer, device=0)
             print("Model Loaded")
+            # run prediction
+            self.generator("My Name is GeniusBot")
+            self.loaded = True
         else:
-            print(f"Device doesn't have at least 14 GB of RAM. Available RAM: {round(float((self.free_memory)/1073741824), 2)}")
+            print(f"Device doesn't have at least {self.memory_required} GB of RAM. "
+                  f"Available RAM: {round(float((self.free_memory)/1073741824), 2)}")
+            self.loaded = False
+
+    def chat(self, text):
+        generated_response = self.generator(text)
+        return generated_response
+
+    def get_loaded(self):
+        return self.loaded
+
 
 if __name__ == "__main__":
     geniusbot_chat = ChatBot()
-    # geniusbot_chat.save_model()
     geniusbot_chat.check_hardware()
+    geniusbot_chat.save_model()
     geniusbot_chat.load_model()
-
-    #geniusbot(sys.argv[1:])
